Connected to #38: Refactore AlertLegend component
